fix(percepta): restore vendored runtime.h, unblock percepta_compile --all-features (Issue 004)
The percepta_compile feature's include_str! pointed at a gitignored
.raw/transformer-vm/ path that was never populated, breaking
cargo check --all-features with a hard error.

Restored the upstream Percepta runtime.h (Apache-2.0, byte-identical to
Percepta-Core/transformer-vm@main) to a tracked location src/percepta/runtime.h
(next to compile.rs) and updated include_str! to the idiomatic "runtime.h".
Moving it out of .raw/ was required because .raw/ is gitignored — restoring
there would only fix local builds, not clones/CI.

Unblocking the module's first-ever compile surfaced ~30 latent masked test bugs
(the include_str! failure had hidden them). Fixed across 6 files:
  - compile.rs/runner.rs/test_percepta_rust_wasm.rs: op == "X" -> *op == "X"
    (DispatchEntry is (&str, i32); iterating yields &&str, so the comparison
    never type-checked once the module actually compiled)
  - cumsum.rs/integration.rs: insert(.., i as i64) -> i as i32 (seq param is i32)
  - bench_064_futamura_evaluator.rs: ProgramGraph::num_dimensions() never
    existed -> all_dims.len(); move-after-match on Runner::evaluate result ->
    match by reference

Verification:
  cargo check --all-features             # EXIT 0 (was: 1 hard error)
  cargo check                            # EXIT 0 (default unaffected)
  cargo test -p katgpt-core --lib        # 509 passed (G3 no-regression)
  cargo test --features percepta_compile --lib -- percepta::
                                         # 339 passed, incl. 4 C->WASM e2e tests
  cargo test --features percepta_compile --test test_percepta_rust_wasm
                                         # 18 passed (Rust->WASM->prefix pipeline)

One pre-existing logic assertion in bench_064
(proof_futamura_specialized_has_fewer_dimensions) remains failing — its premise
that specialization reduces all_dims.len() is wrong (Futamura trades generic
fetch attention for per-instruction dimensions, so dim count grows). Documented
in the issue; not weakened since the author's intended metric is unclear.

+++ src/percepta/cumsum.rs
@@ -116,7 +116,7 @@ mod tests {
 
         for i in 1..=5 {
             let val = i as f64;
-            cs.insert(val, i as f64, i as i64);
+            cs.insert(val, i as f64, i as i32);
             expected += val;
             let got = cs.query();
             assert!(
@@ -137,7 +137,7 @@ mod tests {
         let mut expected = 0.0_f64;
 
         for (i, &val) in fibs.iter().enumerate() {
-            cs.insert(val, i as f64, i as i64);
+            cs.insert(val, i as f64, i as i32);
             expected += val;
             let got = cs.query();
             assert!(
